// src/enums/products.enum.ts
// --- Products Categories Enum
export enum categoryType {
  operatingSystem = 'Operating System',
  applicationSoftware = 'Application Software'
}


// --- Products Platform Type Enum
export enum platformType {
  windows = 'windows',
  mac = 'mac',
  linux = 'linux',
  android = 'android',
  ios = 'ios'
}


// --- Products Base Type Enum
export enum baseType {
  computer = 'computer',
  mobile = 'mobile'
}